+++ site/src/pages/sitemap.xml.ts
@@ -1,0 +1,16 @@
+export const prerender = true;
+
+export function GET() {
+  return new Response(
+    `<?xml version="1.0" encoding="UTF-8"?>
+<urlset xmlns="http://www.sitemaps.org/schemas/sitemap/0.9">
+  <url>
+    <loc>https://shadowpath.coriolislabs.ca/</loc>
+    <changefreq>weekly</changefreq>
+    <priority>1.0</priority>
+  </url>
+</urlset>
+`,
+    { headers: { "content-type": "application/xml; charset=utf-8" } },
+  );
+}